Migrate completion report task to API V1. Add new plugin settings. Add new backend definitions. Remove unnecessary backend definitions.
