--- openedx_pearson_reports/edxapp_wrapper/backends/completion_models_g_v1.py
@@ -1,2 +1,0 @@
-""" Backend abstraction. """
-from lms.djangoapps.completion.models import BlockCompletion  # pylint: disable=unused-import
