+++ openedx_pearson_reports/edxapp_wrapper/backends/course_api_j_v1.py
@@ -1,0 +1,2 @@
+""" Backend abstraction """
+from lms.djangoapps.course_api.blocks.serializers import SUPPORTED_FIELDS  # pylint: disable=unused-import
