--- openedx_pearson_reports/edxapp_wrapper/backends/supported_fields_i_v1.py
@@ -1,2 +1,0 @@
-""" Backend abstraction """
-from lms.djangoapps.course_api.blocks.transformers import SUPPORTED_FIELDS  # pylint: disable=unused-import
