+++ openedx_pearson_reports/edxapp_wrapper/get_course_api.py
@@ -6,7 +6,7 @@ from django.conf import settings
 def get_supported_fields(*args, **kwargs):  # pylint: disable=unused-argument
     """ Get SUPPORTED FIELDS """
 
-    backend_function = settings.OPR_SUPPORTED_FIELDS
+    backend_function = settings.OPR_COURSE_API
     backend = import_module(backend_function)
 
     return backend.SUPPORTED_FIELDS
